leader_cargo/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
Django==3.2.20
django-debug-toolbar==4.4.2
django-htmx==1.17.0
djangorestframework==3.15.1
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
openpyxl==3.1.2
Pillow==10.0.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytweening==1.0.7
pytz==2023.3
pywin32==306
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.0.7
Werkzeug==3.0.1
wikipedia==1.4.0
xlrd==2.0.1
XlsxWriter==3.2.0
xlwings==0.30.12
